[APP-5041] typescript QuestionCard component (#1231)
